GIS/geoai_segmentation/requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.0
torchvision==0.16.1
tensorboard==2.14.0
tqdm==4.66.1
scipy==1.11.3
gdal==3.6.4
rasterio==1.3.9
geopandas==0.14.0
scikit-image==0.22.0